Tidy after refactor
