+++ libsq/core/ioz/linesplitreaders/linesplitreaders_test.go
@@ -27,50 +27,56 @@ var testCases = []struct {
 	{
 		name:         "a-lf",
 		in:           "a\n",
-		wantLines:    []string{"a"},
-		wantRdrCount: 1,
+		wantLines:    []string{"a", ""},
+		wantRdrCount: 2,
 	},
 	{
 		name:         "empty-1-lf",
 		in:           "\n",
-		wantLines:    []string{""},
-		wantRdrCount: 1,
+		wantLines:    []string{"", ""},
+		wantRdrCount: 2,
 	},
 	{
 		name:         "empty-1-cr",
 		in:           "\r",
-		wantLines:    []string{""}, // REVISIT: What do we really want here?
+		wantLines:    []string{"\r"},
 		wantRdrCount: 1,
 	},
 	{
 		name:         "empty-1-crlf",
 		in:           "\r\n",
-		wantLines:    []string{""},
-		wantRdrCount: 1,
+		wantLines:    []string{"", ""},
+		wantRdrCount: 2,
 	},
 	{
 		name:         "empty-2-crlf",
 		in:           "\r\n\r\n",
-		wantLines:    []string{"", ""},
-		wantRdrCount: 2,
+		wantLines:    []string{"", "", ""},
+		wantRdrCount: 3,
 	},
 	{
 		name:         "a-crlf",
 		in:           "a\r\n",
-		wantLines:    []string{"a"},
-		wantRdrCount: 1,
+		wantLines:    []string{"a", ""},
+		wantRdrCount: 2,
+	},
+	{
+		name:         "a-crlf-b-crlf",
+		in:           "a\r\nb\r\n",
+		wantLines:    []string{"a", "b", ""},
+		wantRdrCount: 3,
 	},
 	{
 		name:         "empty-2-lf",
 		in:           "\n\n",
-		wantLines:    []string{"", ""},
-		wantRdrCount: 2,
+		wantLines:    []string{"", "", ""},
+		wantRdrCount: 3,
 	},
 	{
 		name:         "oneline-crlf",
 		in:           "line1\r\n",
-		wantLines:    []string{"line1"},
-		wantRdrCount: 1,
+		wantLines:    []string{"line1", ""},
+		wantRdrCount: 2,
 	},
 	{
 		name:         "oneline-no-lf",
@@ -81,20 +87,20 @@ var testCases = []struct {
 	{
 		name:         "content-2-lf",
 		in:           "line1\nline2\n",
-		wantLines:    []string{"line1", "line2"},
-		wantRdrCount: 2,
+		wantLines:    []string{"line1", "line2", ""},
+		wantRdrCount: 3,
 	},
 	{
 		name:         "ab-2-lf",
 		in:           "ab\ncd\n",
-		wantLines:    []string{"ab", "cd"},
-		wantRdrCount: 2,
+		wantLines:    []string{"ab", "cd", ""},
+		wantRdrCount: 3,
 	},
 	{
 		name:         "ab-crlf-cd-crlf",
 		in:           "ab\r\ncd\r\n",
-		wantLines:    []string{"ab", "cd"},
-		wantRdrCount: 2,
+		wantLines:    []string{"ab", "cd", ""},
+		wantRdrCount: 3,
 	},
 	{
 		name:         "content-4-no-trailing-lf",
@@ -120,17 +126,12 @@ var testCases = []struct {
 		wantLines:    []string{"a", "b", "c", "d"},
 		wantRdrCount: 4,
 	},
-	// NOTE: The below test is commented out, because it's not entirely clear what
-	// we're hoping for here. Either which way, the code is broken, because
-	// different results are returned depending on the size of the []byte passed
-	// to Read.
-	//
-	// {
-	//	name:         "single-char-4-cr",
-	//	in:           "a\rb\rc\rd",
-	//	wantLines:    []string{"abcd"},
-	//	wantRdrCount: 1,
-	// },
+	{
+		name:         "single-char-4-cr",
+		in:           "a\rb\rc\rd",
+		wantLines:    []string{"a\rb\rc\rd"},
+		wantRdrCount: 1,
+	},
 	{
 		name:         "multi-lines-with-extra-lf",
 		in:           "\nline2\nline3\nline4\n\nline5",
@@ -158,8 +159,8 @@ var testCases = []struct {
 	{
 		name:         "single-char-lines-with-extra-lf-2",
 		in:           "\nb\nc\nd\n\nf\n",
-		wantLines:    []string{"", "b", "c", "d", "", "f"},
-		wantRdrCount: 6,
+		wantLines:    []string{"", "b", "c", "d", "", "f", ""},
+		wantRdrCount: 7,
 	},
 	{
 		name:         "a-c-lines-with-extra-lf",
@@ -173,11 +174,18 @@ var testCases = []struct {
 		wantLines:    []string{"a", "", "c"},
 		wantRdrCount: 3,
 	},
+
 	{
 		name:         "lf-lf-c",
 		in:           "\n\nc",
 		wantLines:    []string{"", "", "c"},
 		wantRdrCount: 3,
+	},
+	{
+		name:         "single-char-2-cr",
+		in:           "a\rb\r",
+		wantLines:    []string{"a\rb\r"},
+		wantRdrCount: 1,
 	},
 	{
 		name:         "crlf-crlf-c",
@@ -212,204 +220,20 @@ func Test_ReadAll(t *testing.T) {
 
 // Test_Reader_Read tests via the io.Reader returned from Splitter.Reader.
 func Test_Reader_Read(t *testing.T) {
-	//t.Parallel()
-
-	var testCases = []struct {
-		name         string
-		in           string
-		wantLines    []string
-		wantRdrCount int
-	}{
-		{
-			name:         "empty",
-			in:           "",
-			wantLines:    []string{""},
-			wantRdrCount: 1,
-		},
-		{
-			name:         "a-lf",
-			in:           "a\n",
-			wantLines:    []string{"a", ""},
-			wantRdrCount: 2,
-		},
-		{
-			name:         "empty-1-lf",
-			in:           "\n",
-			wantLines:    []string{"", ""},
-			wantRdrCount: 2,
-		},
-		{
-			name:         "empty-1-cr",
-			in:           "\r",
-			wantLines:    []string{"\r"},
-			wantRdrCount: 1,
-		},
-		{
-			name:         "empty-1-crlf",
-			in:           "\r\n",
-			wantLines:    []string{"", ""},
-			wantRdrCount: 2,
-		},
-		{
-			name:         "empty-2-crlf",
-			in:           "\r\n\r\n",
-			wantLines:    []string{"", "", ""},
-			wantRdrCount: 3,
-		},
-		{
-			name:         "a-crlf",
-			in:           "a\r\n",
-			wantLines:    []string{"a", ""},
-			wantRdrCount: 2,
-		},
-		{
-			name:         "a-crlf-b-crlf",
-			in:           "a\r\nb\r\n",
-			wantLines:    []string{"a", "b", ""},
-			wantRdrCount: 3,
-		},
-		{
-			name:         "empty-2-lf",
-			in:           "\n\n",
-			wantLines:    []string{"", "", ""},
-			wantRdrCount: 3,
-		},
-		{
-			name:         "oneline-crlf",
-			in:           "line1\r\n",
-			wantLines:    []string{"line1", ""},
-			wantRdrCount: 2,
-		},
-		{
-			name:         "oneline-no-lf",
-			in:           "line1",
-			wantLines:    []string{"line1"},
-			wantRdrCount: 1,
-		},
-		{
-			name:         "content-2-lf",
-			in:           "line1\nline2\n",
-			wantLines:    []string{"line1", "line2", ""},
-			wantRdrCount: 3,
-		},
-		{
-			name:         "ab-2-lf",
-			in:           "ab\ncd\n",
-			wantLines:    []string{"ab", "cd", ""},
-			wantRdrCount: 3,
-		},
-		{
-			name:         "ab-crlf-cd-crlf",
-			in:           "ab\r\ncd\r\n",
-			wantLines:    []string{"ab", "cd", ""},
-			wantRdrCount: 3,
-		},
-		{
-			name:         "content-4-no-trailing-lf",
-			in:           "line1\nline2\nline3\nline4",
-			wantLines:    []string{"line1", "line2", "line3", "line4"},
-			wantRdrCount: 4,
-		},
-		{
-			name:         "content-4-no-trailing-crlf",
-			in:           "line1\r\nline2\r\nline3\r\nline4",
-			wantLines:    []string{"line1", "line2", "line3", "line4"},
-			wantRdrCount: 4,
-		},
-		{
-			name:         "single-char-4-lf",
-			in:           "a\nb\nc\nd",
-			wantLines:    []string{"a", "b", "c", "d"},
-			wantRdrCount: 4,
-		},
-		{
-			name:         "single-char-4-crlf",
-			in:           "a\r\nb\r\nc\r\nd",
-			wantLines:    []string{"a", "b", "c", "d"},
-			wantRdrCount: 4,
-		},
-		{
-			name:         "single-char-4-cr",
-			in:           "a\rb\rc\rd",
-			wantLines:    []string{"a\rb\rc\rd"},
-			wantRdrCount: 1,
-		},
-		{
-			name:         "multi-lines-with-extra-lf",
-			in:           "\nline2\nline3\nline4\n\nline5",
-			wantLines:    []string{"", "line2", "line3", "line4", "", "line5"},
-			wantRdrCount: 6,
-		},
-		{
-			name:         "multi-lines-with-extra-crlf",
-			in:           "\r\nline2\r\nline3\r\nline4\r\n\r\nline5",
-			wantLines:    []string{"", "line2", "line3", "line4", "", "line5"},
-			wantRdrCount: 6,
-		},
-		{
-			name:         "single-char-lines-with-extra-lf",
-			in:           "\nb\nc\nd\n\nf",
-			wantLines:    []string{"", "b", "c", "d", "", "f"},
-			wantRdrCount: 6,
-		},
-		{
-			name:         "single-char-lines-with-extra-crlf",
-			in:           "\r\nb\r\nc\r\nd\r\n\r\nf",
-			wantLines:    []string{"", "b", "c", "d", "", "f"},
-			wantRdrCount: 6,
-		},
-		{
-			name:         "single-char-lines-with-extra-lf-2",
-			in:           "\nb\nc\nd\n\nf\n",
-			wantLines:    []string{"", "b", "c", "d", "", "f", ""},
-			wantRdrCount: 7,
-		},
-		{
-			name:         "a-c-lines-with-extra-lf",
-			in:           "a\n\nc",
-			wantLines:    []string{"a", "", "c"},
-			wantRdrCount: 3,
-		},
-		{
-			name:         "a-c-lines-with-extra-crlf",
-			in:           "a\r\n\r\nc",
-			wantLines:    []string{"a", "", "c"},
-			wantRdrCount: 3,
-		},
-
-		{
-			name:         "lf-lf-c",
-			in:           "\n\nc",
-			wantLines:    []string{"", "", "c"},
-			wantRdrCount: 3,
-		},
-		{
-			name:         "single-char-2-cr",
-			in:           "a\rb\r",
-			wantLines:    []string{"a\rb\r"},
-			wantRdrCount: 1,
-		},
-		{
-			name:         "crlf-crlf-c",
-			in:           "\r\n\r\nc",
-			wantLines:    []string{"", "", "c"},
-			wantRdrCount: 3,
-		},
-	}
+	// t.Parallel()
 
 	const bufMin, bufMax = 1, 11
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			//t.Parallel()
+			// t.Parallel()
 			for bufSize := bufMin; bufSize <= bufMax; bufSize++ {
 				t.Run(fmt.Sprintf("buf-%d", bufSize), func(t *testing.T) {
-
 					t.Logf("\n\n>%s<\n\n", tc.in)
 
-					//t.Parallel()
+					// t.Parallel()
 					rdrCount := 0
-					splitter := linesplitreaders.New2(strings.NewReader(tc.in))
+					splitter := linesplitreaders.New(strings.NewReader(tc.in))
 					var lines []string
 
 					for splitter.Next() {
